Use short-circuiting boolean operators in conditions when possible.
This is a hack, it is not desirable for everyone and should probably
be behind an option and disabled by default.

Also the node expansion pass is certainly not the proper place (maybe
introduce a pre-pass dedicated to this kind of transformations).

Side note: it would probably be a bad idea to do this transformation
in the main conversion pass because the code which handles the
short-circuiting operators in very different in the nullable case.
Also it just seems better to separate concerns and keep the main
conversion pass as simple as possible.